refactor: separate export for browser
